@@ -1,22 +1,13 @@
 import { DatabaseQuery, ChildEvent, AngularFireAction, SnapshotAction } from '../interfaces';
 import { fromRef } from '../observable/fromRef';
 import { validateEventsArray } from './utils';
-import { Observable } from 'rxjs';
-import { merge } from 'rxjs/observable/merge';
+import { Observable, merge } from 'rxjs';
 
-import { DataSnapshot } from '@firebase/database-types';
+import { DatabaseSnapshot } from '../interfaces';
 import { AngularFireDatabase } from '../database';
 
-export function createStateChanges(query: DatabaseQuery, afDatabase: AngularFireDatabase) {
-  return (events?: ChildEvent[]) => afDatabase.scheduler.keepUnstableUntilFirst(
-    afDatabase.scheduler.runOutsideAngular(
-      stateChanges(query, events)
-    )
-  );
-}
-
-export function stateChanges(query: DatabaseQuery, events?: ChildEvent[]) {
+export function stateChanges<T>(query: DatabaseQuery, events?: ChildEvent[]) {
   events = validateEventsArray(events)!;
-  const childEvent$ = events.map(event => fromRef(query, event));
+  const childEvent$ = events.map(event => fromRef<T>(query, event));
   return merge(...childEvent$);
 }
