@@ -3,11 +3,11 @@ import { DatabaseQuery, AngularFireObject } from '../interfaces';
 import { createObjectSnapshotChanges } from './snapshot-changes';
 import { AngularFireDatabase } from '../database';
 
-export function createObjectReference<T>(query: DatabaseQuery, afDatabase: AngularFireDatabase): AngularFireObject<T> {
+export function createObjectReference<T=any>(query: DatabaseQuery, afDatabase: AngularFireDatabase): AngularFireObject<T> {
   return {
     query,
     snapshotChanges<T>() {
-      const snapshotChanges$ = createObjectSnapshotChanges(query)();
+      const snapshotChanges$ = createObjectSnapshotChanges<T>(query)();
       return afDatabase.scheduler.keepUnstableUntilFirst(
         afDatabase.scheduler.runOutsideAngular(
           snapshotChanges$
